Split adapter behavior into reusable mixins
We'll want to use the methods that create the uath domain and that
share a workspace with the admins elsewhere, so move the code into
a set of resuable classes that can be inherited in the workspae
adapters.
