+++ primed/cdsa/adapters.py
@@ -1,14 +1,14 @@
 from anvil_consortium_manager.adapters.workspace import BaseWorkspaceAdapter
-from anvil_consortium_manager.models import GroupGroupMembership, ManagedGroup, Workspace, WorkspaceGroupSharing
-from django.conf import settings
+from anvil_consortium_manager.models import Workspace
 
 from primed.miscellaneous_workspaces.tables import DataPrepWorkspaceUserTable
+from primed.primed_anvil.adapters import WorkspaceAdminSharingAdapter, WorkspaceAuthDomainAdapter
 from primed.primed_anvil.forms import WorkspaceAuthDomainDisabledForm
 
 from . import forms, models, tables
 
 
-class CDSAWorkspaceAdapter(BaseWorkspaceAdapter):
+class CDSAWorkspaceAdapter(WorkspaceAuthDomainAdapter, WorkspaceAdminSharingAdapter, BaseWorkspaceAdapter):
     """Adapter for CDSAWorkspaces."""
 
     type = "cdsa"
@@ -33,34 +33,3 @@ class CDSAWorkspaceAdapter(BaseWorkspaceAdapter):
             extra_context["primary_cdsa"] = None
 
         return extra_context
-
-    def before_anvil_create(self, workspace):
-        # Create the auth domain for the workspace.
-        """Add authorization domain to workspace."""
-        auth_domain_name = "AUTH_" + workspace.name
-        auth_domain = ManagedGroup.objects.create(
-            name=auth_domain_name,
-            is_managed_by_app=True,
-            email=auth_domain_name + "@firecloud.org",
-        )
-        workspace.authorization_domains.add(auth_domain)
-        auth_domain.anvil_create()
-        # Add the ADMINs group as an admin of the auth domain.
-        admins_group = ManagedGroup.objects.get(name=settings.ANVIL_CC_ADMINS_GROUP_NAME)
-        membership = GroupGroupMembership.objects.create(
-            parent_group=auth_domain,
-            child_group=admins_group,
-            role=GroupGroupMembership.ADMIN,
-        )
-        membership.anvil_create()
-
-    def after_anvil_create(self, workspace):
-        # Share the workspace with the ADMINs group as an owner.
-        admins_group = ManagedGroup.objects.get(name=settings.ANVIL_CC_ADMINS_GROUP_NAME)
-        sharing = WorkspaceGroupSharing.objects.create(
-            workspace=workspace,
-            group=admins_group,
-            access=WorkspaceGroupSharing.OWNER,
-            can_compute=True,
-        )
-        sharing.anvil_create_or_update()
